refactor: split fonts helpers and tab panel
